+++ src/workers/validator_worker.js
